refactor: Изменение нейминга реквестов, удаление дублирующихся реквестов
